kristianaspevik/requirements.txt
Django==2.0
gunicorn==19.9.0
pytz==2018.5
djangorestframework==3.10.2
